# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User Management API",
        "description": "API documentation for managing users",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Local server"
        }
    ],
    "paths": {
        "/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of users",
                "description": "Returns list of users with optional status filter",
                "operationId": "23679253d16f34d2972249751e3631dd",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter users by status (0 = inactive, 1 = active, 2 = deleted)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "operationId": "ce3cf31668a69eb39e5e98738715afa1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone_number",
                                    "email",
                                    "status",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "status": {
                                        "type": "integer"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/user/create": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Show form to create a new user",
                "operationId": "8c02c9f610b5c714fa4921bb178f92a6",
                "responses": {
                    "200": {
                        "description": "Form returned"
                    }
                }
            }
        },
        "/user/{id}/edit": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Show form to edit a user",
                "operationId": "bec114dd70fcef1a67d9bf1b48cc2036",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user to edit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User data returned"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/user/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update a user",
                "operationId": "786a794071cb7098f55a4a4911e7d3f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone_number",
                                    "email",
                                    "status",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "status": {
                                        "type": "integer"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a user",
                "operationId": "a56b12a3694a926699892056cdcf8a9e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/user/destroy-bulk": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Bulk delete users",
                "operationId": "1d2ad75bfe3c6ab8e9370252aadce078",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Users deleted successfully"
                    },
                    "422": {
                        "description": "No users selected"
                    }
                }
            }
        },
        "/user/export": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Export users to Excel",
                "operationId": "2447f8f72321bfd3c57ff033a41fd4b3",
                "responses": {
                    "200": {
                        "description": "Excel file download"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "required": [
                    "id",
                    "name",
                    "email",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "+60123456789"
                    },
                    "status": {
                        "description": "0=inactive,1=active,2=deleted",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Users",
            "description": "API Endpoints for managing users"
        }
    ]
}